input.c: deleted, functions reorganized to encoding.c/decoding.c

+++ lab5/src/headers/typedefs.h
@@ -4,7 +4,7 @@
 #include "stdint.h"
 
 typedef unsigned char uchar;
-typedef uint8_t byte;
+typedef unsigned int uint;
 
 typedef struct TreeNode TreeNode;
 typedef struct CharInfo CharInfo;
